Telegram refactoring.
