+++ src/Telegram/BotModel.hs
@@ -1,0 +1,8 @@
+module Telegram.BotModel where
+
+type LastUpdateId = Int
+
+type ChatId = Int
+
+botUri :: String
+botUri = "https://api.telegram.org/bot"
